One more step.

+++ tasks/ghPages.js
@@ -23,58 +23,50 @@ module.exports = function (grunt) {
 	// Please see the Grunt documentation for more information regarding task
 	// creation: http://gruntjs.com/creating-tasks
 
-	grunt.registerMultiTask('ghPages', 'Grunt plugin for easy deployment to ghPages branch.', function () {
+	grunt.registerTask('ghPages', 'Grunt plugin for easy deployment to ghPages branch.', function () {
 
 		// Merge task-specific and/or target-specific options with these defaults.
 		var options = this.options({
-			repository: '',
-			branch: 'gh-pages'
+			repository: null,
+			branch: 'gh-pages',
+			deployPath: null,
+			message: 'Deployment ' + grunt.template.today()
 		});
 
-		console.log('ghPages task', options);
+		if (!options.repository) {
+			grunt.log.error('Repository is required.');
+			return false;
+		}
+
+		if (!options.deployPath) {
+			grunt.log.error('Deployment path is required.');
+			return false;
+		}
+
+		if (!grunt.file.isDir(options.deployPath)) {
+			grunt.log.error('Deployment path "' + options.deployPath + '" is not directory. Nothing to deploy.');
+			return false;
+		}
 
 		var command = [
+			'rm -rf .tmp-ghpages',
 			'mkdir .tmp-ghpages',
 			'cd .tmp-ghpages',
 			'git init',
-			'git remote add -t ' + options.branch + ' -f origin ' + options.repository,
+			'git remote add -t ' + options.branch + ' -f origin ../' + options.repository,
 			'git checkout ' + options.branch,
 			'find -not -path "./.git/*" -not -name ".git" -delete',
 			'cp -r ../' + options.deployPath + '/* ./',
 			'git add -A',
-			'git commit -m "Some message"',
+			'git commit -m "' + options.message + '"',
 			'git push origin ' + options.branch
 		].join(' && ');
 
 		console.log(command);
+		runCmd(command, function() {
+			console.log('Result', arguments);
+		});
 
-
-
-		// Iterate over all specified file groups.
-//		this.files.forEach(function (f) {
-//			// Concat specified files.
-//			var src = f.src.filter(function (filepath) {
-//				// Warn on and remove invalid source files (if nonull was set).
-//				if (!grunt.file.exists(filepath)) {
-//					grunt.log.warn('Source file "' + filepath + '" not found.');
-//					return false;
-//				} else {
-//					return true;
-//				}
-//			}).map(function (filepath) {
-//				// Read file source.
-//				return grunt.file.read(filepath);
-//			}).join(grunt.util.normalizelf(options.separator));
-//
-//			// Handle options.
-//			src += options.punctuation;
-//
-//			// Write the destination file.
-//			grunt.file.write(f.dest, src);
-//
-//			// Print a success message.
-//			grunt.log.writeln('File "' + f.dest + '" created.');
-//		});
 	});
 
 };
